dplcompat: allow driver-specific configuration
with this change the program that will be used to access the storage can
declare what options it supports (via the "options" subcommand).
If one of those options is in Device Options, it will be added to the
program's environment.
If there are options in Device Options that neither dplcompat nor the
program accept, this is treated as a configuration error.